+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+- Enable passive mode in addition to active mode
 
 ## [0.1.0] - 2025-07-02
 - Initial version